requirements.txt
torch
torchvision
streamlit
tqdm
Pillow
pytest